refactor: 전체 시험 목록과 내 시험 목록 interface 분리
